rename: userRole 폴더 분리
