+++ src/auth/member.entity.ts
@@ -14,7 +14,7 @@ import {
   Unique,
   UpdateDateColumn,
 } from 'typeorm';
-import { UserRole } from './dto/auth-credential.dto';
+import { UserRole } from './role/user.role';
 
 @Entity()
 @Unique(['email'])
@@ -94,13 +94,13 @@ export class member extends BaseEntity {
   @JoinColumn()
   semesters: semester[];
 
-  @OneToMany(() => course, (course) => course.memberId, {
+  @OneToMany(() => course, (course) => course, {
     eager: false,
   })
   @JoinColumn()
   courses: course[];
 
-  @OneToMany(() => attendance, (attendance) => attendance.memberId, {
+  @OneToMany(() => attendance, (attendance) => attendance, {
     eager: false,
   })
   @JoinColumn()
