+++ src/auth/role/user.role.ts
@@ -1,0 +1,4 @@
+export enum UserRole {
+  MEMBER = 'MEMBER', // index 0
+  MANAGER = 'MANAGER', // index 1
+}
